+++ src/onegov.town/onegov/town/models/__init__.py
@@ -1,12 +1,13 @@
 from onegov.town.models.atoz import AtoZ
+from onegov.town.models.auth import Auth
 from onegov.town.models.clipboard import Clipboard
 from onegov.town.models.editor import Editor
 from onegov.town.models.file import File, FileCollection
 from onegov.town.models.form import BuiltinFormDefinition, CustomFormDefinition
 from onegov.town.models.image import Image, ImageCollection, Thumbnail
 from onegov.town.models.page import News, Topic
-from onegov.town.models.search import Search
 from onegov.town.models.resource import DaypassResource
+from onegov.town.models.search import Search
 from onegov.town.models.sitecollection import SiteCollection
 from onegov.town.models.town import Town
 from onegov.town.models.traitinfo import TraitInfo
@@ -14,6 +15,7 @@ from onegov.town.models.traitinfo import TraitInfo
 
 __all__ = [
     'AtoZ',
+    'Auth',
     'BuiltinFormDefinition',
     'Clipboard',
     'CustomFormDefinition',